(1).取消部分mxic敘述.(2).更新PageModelBase的首頁來自appSettings.(3).加入EDMSqlServer套件參考，與範本對SqlServer的支援

+++ EasyArchitect.OutsideManaged.AuthExtensions/Models/Accountvo.cs
@@ -10,6 +10,6 @@ namespace EasyArchitect.OutsideManaged.AuthExtensions.Models
         /// <summary>
         /// 流水號（識別用）
         /// </summary>
-        public decimal? ID { get; set; }
+        public int? ID { get; set; }
     }
 }
